[OS-0] Fixed TDS

+++ app/src/main/java/com/delhivery/axle/api/response/PayableResponse.kt
@@ -68,7 +68,7 @@ data class TDS(
             ) {
                 return (amount * (updatedTDSRate / 100))
             } else {
-                return (amount * (tdsRate / 100))
+                return (amount * (tdsRate.toDouble() / 100))
             }
         } else {
             return 0.0
